Merged PR 85: Add blob garbage collection
Creating a staging table in the metadata database. The create flow is now:
1. Write a staged record in the metadata database
2. Write the blob to the blob store,
3. "Complete" the staged metadata record.

If (2) fails, we delete the staged record.

For cases where there is a crash between (2) and (3), we have a simple garbage collection routine, which pulls staging records that are older than 30 minutes and deletes their corresponding blob and the staging record. We repeat this process every 30 minutes.